This CL is for the following.... no wait.... me before the CL. >_>
So I had a haircut today. But did you notice? No you did not and now you have to deal with my petulance.
Who ever knew that writing code for computing BLEU isnt so straightforward? I learned the hard way. I also learned the actual algorithm for BLEU. Been in MT for like 4 years and knew that it was N-gram precision but never actually bothered to implement. Fancy that! Atleast now I know that some of the things dont work exactly as I imagined they did. Then came the tests. Boy they took a bit of time. Also, me and a bunch of friends are doing Hanami on Thursday. That will be fun it will! Finally, did anyone check the trailer for Thor: Ragnarok? Jesus Christ! They put the "Immigrant Song" by Led Zepplin in it and I have been listening to the trailer for over 5 hours. Some songs are gold! Now back to the CL!

This CL is about the following:
1. An implementation of BLEU for MT evaluation so that I can generate some debug information.
2. Tests for bleu generation unver various circumstances. Should I add more? Not really sure. This stuff seems to work exactly as it is supposed to.